requirements/common.txt
amqp==5.2.0
appnope==0.1.3
asgiref==3.7.2
backports.zoneinfo;python_version<"3.9"
beautifulsoup4==4.12.2
billiard==4.2.0
boto3==1.34.39
botocore==1.34.39
celery==5.3.5
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.11
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-htmlmin==0.11.0
django-impersonate==1.9.1
djangorestframework==3.14.0
html5lib==1.1
jmespath==1.0.1
kombu==5.3.3
numpy==1.24.4
pandas==2.0.3
phonenumbers==8.13.23
pip-autoremove==0.10.0
pipdeptree==2.13.0
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.10.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.4
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.8
webencodings==0.5.1
# pygraphviz==1.11 --config-settings="--global-option=build_ext" --config-settings="--global-option=-I$(brew --prefix graphviz)/include/" --config-settings="--global-option=-L$(brew --prefix graphviz)/lib/"; sys_platform == "darwin"
